% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.default}
\alias{as.data.frame.default}
\title{S3 support}
\usage{
as.data.frame.default(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{entity coercible to data.frame}

\item{row.names}{character or NULL}

\item{optional}{logical}

\item{\dots}{any args}
}
\description{
S3 support
}
