% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.hail.table.Table}
\alias{as.data.frame.hail.table.Table}
\title{convert hail.table.Table to R data frame}
\usage{
\method{as.data.frame}{hail.table.Table}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{instance of "hail.table.Table"}

\item{row.names}{not used}

\item{optional}{not used}

\item{\dots}{not used}
}
\value{
data.frame
}
\description{
convert hail.table.Table to R data frame
}
\note{
only use on small table because collect is used
}
\examples{
hl <- hail_init()
annopath <- path_1kg_annotations()
tab <- hl$import_table(annopath, impute = TRUE)$key_by("Sample")
as.data.frame(tab$head(3L))
}
