% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail_control.R
\name{bare_hail}
\alias{bare_hail}
\title{bare interface to hail using reticulate}
\usage{
bare_hail()
}
\description{
bare interface to hail using reticulate
}
\examples{
# assumes terra
if (nchar(Sys.getenv("WORKSPACE_NAMESPACE"))>0) {
  hl = bare_hail()
  hl$init(idempotent=TRUE, spark_conf=list(
      'spark.hadoop.fs.gs.requester.pays.mode'= 'CUSTOM',
      'spark.hadoop.fs.gs.requester.pays.buckets'= 'ukb-diverse-pops-public',
      'spark.hadoop.fs.gs.requester.pays.project.id'= Sys.getenv("GOOGLE_PROJECT")))
  hl$read_matrix_table('gs://ukb-diverse-pops-public/sumstats_release/results_full.mt')$describe()
  }
}
