% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames.R
\name{colnames.hail.table.Table}
\alias{colnames.hail.table.Table}
\title{acquire column names of a Hail Table}
\usage{
\method{colnames}{hail.table.Table}(x, ...)
}
\arguments{
\item{x}{instance of hail.table.Table}
}
\value{
character()
}
\description{
acquire column names of a Hail Table
}
\note{
writes one line of table to disk to retrieve field names
}
