% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames.R
\name{colnames}
\alias{colnames}
\title{S3 support}
\usage{
colnames(x, do.NULL = TRUE, prefix = "col")
}
\arguments{
\item{x}{array instance}

\item{do.NULL}{logical}

\item{prefix}{character(1)}
}
\description{
S3 support
}
