% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_filter.R
\name{filter}
\alias{filter}
\title{S3 support}
\usage{
filter(
  x,
  filter,
  method = c("convolution", "recursive"),
  sides = 2L,
  circular = FALSE,
  init = NULL,
  ...,
  hl = NULL,
  placeholder = NULL,
  drop_placeholder = NULL
)
}
\description{
S3 support
}
