% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames.R
\name{get_key.hail.table.Table}
\alias{get_key.hail.table.Table}
\title{S3 method for get_key}
\usage{
\method{get_key}{hail.table.Table}(x)
}
\arguments{
\item{x}{instance of hail.table.Table}
}
\value{
a list with elements names (names of keys) and key_df (data.frame of key values, with column names)
}
\description{
S3 method for get_key
}
