% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames.R
\name{get_key}
\alias{get_key}
\title{S3 generic for get_key}
\usage{
get_key(x)
}
\arguments{
\item{x}{anything}
}
\value{
typically a list
}
\description{
S3 generic for get_key
}
\examples{
hl <- hail_init()
annopath <- path_1kg_annotations()
tab <- hl$import_table(annopath, impute = TRUE)$key_by("Sample")
get_key(tab)
}
