% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail_control.R
\name{hail_init_simple}
\alias{hail_init_simple}
\title{initialize hail}
\usage{
hail_init_simple()
}
\description{
initialize hail
}
\note{
hail object may be passed around
}
\examples{
hc <- hail_init_simple()
hc
}
