% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail_control.R
\name{hail_stop}
\alias{hail_stop}
\title{stop hail}
\usage{
hail_stop(hl)
}
\arguments{
\item{hl}{a hail object produced by hail_init()}
}
\value{
result of stop() method for Hail module
}
\description{
stop hail
}
