% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kg_3202}
\alias{kg_3202}
\title{data.frame with metadata about 3202 samples genotyped against T2T reference}
\format{
data.frame
}
\usage{
<<<<<<< HEAD
data('kg_3202')
=======
data("kg_3202")
>>>>>>> 49d1b74... adding colnames and get_key operations for hail.table.Table
}
\description{
data.frame with metadata about 3202 samples genotyped against T2T reference
}
\note{
Source: index files described at `https://www.internationalgenome.org/data-portal/data-collection/30x-grch38`
}
\keyword{datasets}
