% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_1kg.R
\name{osn_1kg_path}
\alias{osn_1kg_path}
\title{Open Storage Network path to a zip of hail MatrixTable with some 1kg data for the Hail.is GWAS tutorial}
\usage{
osn_1kg_path()
}
\description{
Open Storage Network path to a zip of hail MatrixTable with some 1kg data for the Hail.is GWAS tutorial
}
