% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ukbb_sumst.R
\name{osn_ukbb_sumst10k_path}
\alias{osn_ukbb_sumst10k_path}
\title{Open Storage Network path to a zip of hail MatrixTable with a small subset of UKBB summary statistics as of 12/25/2022}
\usage{
osn_ukbb_sumst10k_path()
}
\description{
Open Storage Network path to a zip of hail MatrixTable with a small subset of UKBB summary statistics as of 12/25/2022
}
