% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcs_38k}
\alias{pcs_38k}
\title{HWE-normalized PCA scores for 3202 thousand-genomes samples genotyped with the
telomere-to-telomere reference}
\format{
data.frame
}
\usage{
data("pcs_38k")
}
\value{
data.frame
}
\description{
HWE-normalized PCA scores for 3202 thousand-genomes samples genotyped with the
telomere-to-telomere reference
}
\note{
The genotypes are from a 1% sample of loci on chr17
}
\examples{
data(pcs_38k)
dim(pcs_38k)
}
\keyword{datasets}
