% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_col_ops.R
\name{pheno_data_sec_2df}
\alias{pheno_data_sec_2df}
\title{pheno_data component harvesting from columns of summary stats MatrixTable}
\usage{
pheno_data_sec_2df(
  m,
  section = 1,
  toget = c("n_cases", "n_controls", "heritability", "pop")
)
}
\arguments{
\item{m}{Struct returned from mt$cols()$collect()}

\item{toget}{character() vector of field names to retrieve}
}
\value{
1 row data.frame
}
\description{
pheno_data component harvesting from columns of summary stats MatrixTable
}
\note{
applies top2df to the pheno_data component of input
}
