% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_update.R
\name{rg_update}
\alias{rg_update}
\title{update the reference genome for a hail instance}
\usage{
rg_update(
  hc,
  init = "GRCh38",
  newjson = system.file("json/t2tAnVIL.json", package = "BiocHail")
)
}
\arguments{
\item{hc}{hail context}

\item{init}{character(1) valid name for a reference genome, defaults to "GRCh38"}

\item{newjson}{character(1) path to a json spec of a reference genome [needs doc]}
}
\value{
a python list; the function is used for its side effect
}
\description{
update the reference genome for a hail instance
}
