% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames.R
\name{rownames.hail.table.Table}
\alias{rownames.hail.table.Table}
\title{acquire row names of a Hail Table}
\usage{
\method{rownames}{hail.table.Table}(x, ...)
}
\arguments{
\item{x}{instance of hail.table.Table}
}
\value{
character()
}
\description{
acquire row names of a Hail Table
}
\note{
writes one line of table to disk to retrieve field names
}
\examples{
hl = hail_init()
annopath <- path_1kg_annotations()
tab <- hl$import_table(annopath, impute=TRUE)$key_by("Sample")
rownames(tab)
}
