% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames.R
\name{rownames}
\alias{rownames}
\title{S3 support}
\usage{
rownames(x, do.NULL = TRUE, prefix = "row")
}
\arguments{
\item{x}{array-like entity}

\item{do.NULL}{logical}

\item{prefix}{character}
}
\description{
S3 support
}
