% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_col_ops.R
\name{top2df}
\alias{top2df}
\title{top-level annotation harvesting from columns of summary statistics MatrixTable}
\usage{
top2df(
  x,
  toget = c("trait_type", "phenocode", "description", "modifier", "coding_description")
)
}
\arguments{
\item{x}{a Struct returned from mt$cols()$collect() -- which can be slow}

\item{toget}{character() vector of field names to retrieve}
}
\value{
1-row data.frame
}
\description{
top-level annotation harvesting from columns of summary statistics MatrixTable
}
\note{
python None are transformed to NA
}
