% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbb_init.R
\name{ukbb_init}
\alias{ukbb_init}
\title{initialize ukbb}
\usage{
ukbb_init()
}
\value{
python module reference
}
\description{
initialize ukbb
}
\note{
ukbb module may be passed around
}
\examples{
ukbb <- ukbb_init()
names(ukbb)
}
