% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.default}
\alias{as.data.frame.default}
\title{S3 support}
\usage{
\method{as.data.frame}{default}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{entity coercible to data.frame}

\item{row.names}{character or NULL}

\item{optional}{logical}

\item{\dots}{any args}
}
\value{
data.frame
}
\description{
S3 support
}
\examples{
hl = hail_init()
annopath <- path_1kg_annotations()
tab <- hl$import_table(annopath, impute=TRUE)$key_by("Sample")
as.data.frame(tab$head(3L))
}
