% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail_control.R
\name{hail_init2}
\alias{hail_init2}
\title{initialize hail, using more options}
\usage{
hail_init2(
  quiet = FALSE,
  min_block_size = 0L,
  branching_factor = 50L,
  default_reference = "GRCh38",
  global_seed = 1234L,
  spark_conf = NULL,
  gcs_requester_pays_configuration = NULL
)
}
\arguments{
\item{quiet}{logical(1) defaults to FALSE}

\item{min_block_size}{integer(1) defaults to 0L}

\item{branching_factor}{integer(1) defaults to 50L}

\item{default_reference}{character(1) defaults to "GRCh38",}

\item{global_seed}{integer(1) defaults to 1234L}

\item{spark_conf}{list, defaults to NULL}

\item{gcs_requester_pays_configuration}{list, defaults to NULL}
}
\description{
initialize hail, using more options
}
\note{
hail object may be passed around.  See hail documentation for details on all args.
}
\examples{
proj = Sys.getenv("GOOGLE_PROJECT")
buck = Sys.getenv("GCS_BUCKET")
if (nchar(buck)>0) {
  conf = list(proj, c(buck))
  hl <- hail_init2(gcs_requester_pays_configuration=conf)
}  
}
