% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_1kg.R
\name{path_1kg_annotations}
\alias{path_1kg_annotations}
\title{generate path to installed annotations file}
\usage{
path_1kg_annotations()
}
\value{
character(1) path to annotations
}
\description{
generate path to installed annotations file
}
\examples{
path_1kg_annotations()
}
