% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcs_38k}
\alias{pcs_38k}
\title{HWE-normalized PCA scores for 3202 thousand-genomes samples genotyped with the 
telomere-to-telomere reference}
\format{
data.frame
}
\usage{
<<<<<<< HEAD
data('pcs_38k')
=======
data("pcs_38k")
>>>>>>> 49d1b74... adding colnames and get_key operations for hail.table.Table
}
\description{
HWE-normalized PCA scores for 3202 thousand-genomes samples genotyped with the 
telomere-to-telomere reference
}
\note{
The genotypes are from a 1% sample of loci on chr17
}
\keyword{datasets}
